# logging: some optimizations