Please see the commit messages for details.

On this benchmark:

```py
import pytest
@pytest.mark.parametrize("x", range(5000))
def test_foo(x): pass
```

Before (only relevant entries shown):

```
============================ 5000 passed in 14.13s =============================
         13465818 function calls (12550967 primitive calls) in 14.571 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
140003/60003 0.109    0.000    2.435    0.000 contextlib.py:108(__enter__)
    80000    0.098    0.000    2.224    0.000 logging.py:618(_runtest_for)
140003/60003 0.143    0.000    2.194    0.000 contextlib.py:117(__exit__)
    80000    0.257    0.000    1.721    0.000 logging.py:627(_runtest_for_main)
    30000    0.046    0.000    0.972    0.000 logging.py:685(pytest_runtest_logreport)
    40000    0.096    0.000    0.711    0.000 logging.py:302(__init__)
    10000    0.014    0.000    0.456    0.000 logging.py:668(pytest_runtest_teardown)
    80000    0.137    0.000    0.434    0.000 logging.py:271(catching_logs)
    10000    0.017    0.000    0.323    0.000 logging.py:673(pytest_runtest_logstart)
    10000    0.014    0.000    0.301    0.000 logging.py:680(pytest_runtest_logfinish)
    10000    0.013    0.000    0.285    0.000 logging.py:658(pytest_runtest_setup)
    10000    0.012    0.000    0.258    0.000 logging.py:663(pytest_runtest_call)
```

After:

```
============================ 5000 passed in 11.86s =============================
         10476806 function calls (9908460 primitive calls) in 12.164 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    35003    0.053    0.000    1.430    0.000 contextlib.py:117(__exit__)
    35003    0.037    0.000    1.155    0.000 contextlib.py:108(__enter__)
    30000    0.143    0.000    0.563    0.000 logging.py:622(_runtest_for)
    30000    0.067    0.000    0.223    0.000 logging.py:275(catching_logs)
    10000    0.020    0.000    0.222    0.000 logging.py:647(pytest_runtest_setup)
    10000    0.024    0.000    0.222    0.000 logging.py:657(pytest_runtest_teardown)
    10000    0.011    0.000    0.184    0.000 logging.py:653(pytest_runtest_call)
    15000    0.006    0.000    0.006    0.000 logging.py:675(pytest_runtest_logreport)
     5000    0.003    0.000    0.003    0.000 logging.py:663(pytest_runtest_logstart)
     5000    0.003    0.000    0.003    0.000 logging.py:670(pytest_runtest_logfinish)
```