@bluetech 2020-05-12
I fixed the test failure, and now it's also faster than the previous version (updated PR description), so it's ready for review.