@bluetech 2020-05-17
Thanks for the reviews @Zac-HD and @thisch!

Upon looking at logging a bit more, and looking at the open issues/PRs, I think that the main change here ("reuse LoggingCaptureHandler") might have some interactions with them, so I decided to defer it for a bit. Instead, I took the other changes from here (and some additional ones) into a new PR #7224. Once/if it's merged, I will submit a few more changes that I have in mind (probably including the reuse optimization, and some changes to the fixture-plugin interaction).